refactor: remove dead install progress streaming
