+++ internal/npm/npm.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/openbootdotdev/openboot/internal/progress"
 	"github.com/openbootdotdev/openboot/internal/ui"
 )
 
@@ -109,16 +108,12 @@ func InstallContext(ctx context.Context, packages []string, dryRun bool) error {
 		return fmt.Errorf("list installed packages: %w", err)
 	}
 
-	var toInstall, alreadyInstalled []string
+	var toInstall []string
 	for _, p := range packages {
 		if !installed[p] {
 			toInstall = append(toInstall, p)
-		} else {
-			alreadyInstalled = append(alreadyInstalled, p)
 		}
 	}
-	// Streaming invariant: skipped packages still produce a terminal event.
-	EmitSkipped(alreadyInstalled)
 
 	skipped := len(packages) - len(toInstall)
 	if skipped > 0 {
@@ -183,29 +178,17 @@ func warnIfNodeVersionTooLow(packages []string) {
 // fails it falls back to sequential per-package installs. Returns the list of
 // package names that could not be installed and any fatal error.
 func installBatchContext(ctx context.Context, toInstall []string) (failed []string, err error) {
-	if streaming() {
-		progressSink.Emit(progress.Event{Phase: progress.PhaseNpm, Status: progress.StepStart, Command: "npm install -g " + strings.Join(toInstall, " ")})
-	}
-
 	args := append([]string{"install", "-g"}, toInstall...)
 	batchOutput, batchErr := runnerCombinedOutputContext(ctx, args...)
 
 	if batchErr == nil {
-		if streaming() {
-			for _, p := range toInstall {
-				progressSink.Emit(progress.Event{Phase: progress.PhaseNpm, Name: p, Status: progress.StepOK})
-			}
-		} else {
-			ui.Success(fmt.Sprintf("  ✔ %d npm packages installed", len(toInstall)))
-		}
+		ui.Success(fmt.Sprintf("  ✔ %d npm packages installed", len(toInstall)))
 		return nil, nil
 	}
 
 	batchError := parseNpmError(string(batchOutput))
-	if !streaming() {
-		ui.Warn(fmt.Sprintf("Batch install failed (%s), falling back to sequential...", batchError))
-		ui.Println()
-	}
+	ui.Warn(fmt.Sprintf("Batch install failed (%s), falling back to sequential...", batchError))
+	ui.Println()
 
 	return installSequentialContext(ctx, toInstall)
 }
@@ -218,49 +201,31 @@ func installSequentialContext(ctx context.Context, toInstall []string) (failed [
 		return nil, fmt.Errorf("list packages after batch: %w", err)
 	}
 
-	var remaining, batchRecovered []string
+	var remaining []string
 	for _, pkg := range toInstall {
 		if !nowInstalled[pkg] {
 			remaining = append(remaining, pkg)
-		} else {
-			batchRecovered = append(batchRecovered, pkg)
-		}
-	}
-	// Packages the failed batch did manage to install must still produce
-	// their terminal event, or a streaming renderer's totals never complete.
-	if streaming() {
-		for _, pkg := range batchRecovered {
-			progressSink.Emit(progress.Event{Phase: progress.PhaseNpm, Name: pkg, Status: progress.StepOK})
 		}
 	}
 
 	if len(remaining) == 0 {
-		if !streaming() {
-			ui.Success("All npm packages already installed after partial batch!")
-		}
+		ui.Success("All npm packages already installed after partial batch!")
 		return nil, nil
 	}
 
-	// bar stays nil when a streaming sink is registered.
-	var bar *ui.StickyProgress
-	if !streaming() {
-		bar = ui.NewStickyProgress(len(remaining))
-		bar.Start()
-	}
+	bar := ui.NewStickyProgress(len(remaining))
+	bar.Start()
 
 	for _, pkg := range remaining {
 		npmStepStart(bar, pkg)
-		start := time.Now()
 		errMsg := installNpmPackageWithRetryContext(ctx, pkg)
-		npmStepDone(bar, pkg, errMsg == "", errMsg, ui.FormatDuration(time.Since(start)))
+		npmStepDone(bar, pkg, errMsg == "", errMsg)
 		if errMsg != "" {
 			failed = append(failed, pkg)
 		}
 	}
 
-	if bar != nil {
-		bar.Finish()
-	}
+	bar.Finish()
 	return failed, nil
 }
